refactor(config): ClientConfig 패키지 구조 변경 및 테스트 수정
- `ClientConfig`를 `roomescape.infrastructure.payment.config` 패키지로 이동
- `TossPaymentClientTest`에서 새로운 패키지 구조 반영
- 테스트 환경 설정(`@RestClientTest`)에서 `ClientConfig` 추가
